requirements.txt
opencv-python
matplotlib
numpy
Pillow
torch
torchvision
tqdm
pyyaml
requests
pandas
seaborn
flask
mysql.connector
pytesseract